refactor repo, seperate viewsand models
