+++ Be Written/Assets.xcassets/AppIcon.appiconset/Contents.json
