+++ Be Written/Models/Note.swift
@@ -1,0 +1,8 @@
+//
+//  Note.swift
+//  Be Written
+//
+//  Created by Jared Webber on 2023-08-21.
+//
+
+import Foundation
