+++ Be Written/Models/Quote.swift
@@ -1,0 +1,13 @@
+//
+//  Quote.swift
+//  Be Written
+//
+//  Created by Jared Webber on 2023-08-21.
+//
+
+import Foundation
+
+struct Quote {
+    var title: String
+    var quote: String
+}
