+++ README.md
@@ -1,1 +1,6 @@
-This has not a lot atm thanks for taking a look Jared.
+## be-written
+
+> This has not a lot atm thanks for taking a look Jared.
+Word Up
+
+Theres a bunch of different code styles for swift, but I was looking at [this repo](https://github.com/Lickability/swift-best-practices/tree/main) for this branch
